+++ src/controls/BranchSpan.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
     import type { StoreRef } from "../messages/StoreRef";
 
-    export let ref: Extract<StoreRef, { type: "LocalBranch" } | { type: "RemoteBranch" }>;
+    export let ref: Extract<StoreRef, { type: "LocalBookmark" } | { type: "RemoteBookmark" }>;
 </script>
 
 <span class="ref">
-    {ref.branch_name}{#if ref.type == "RemoteBranch"}@{ref.remote_name}{/if}
+    {ref.branch_name}{#if ref.type == "RemoteBookmark"}@{ref.remote_name}{/if}
 </span>
 
 <style>